[Refactor]: Refactor the file synchronizer to load .gitignore & patial sync
